Renamed controller file.
